Record and record pattern

+++ Chapter04/P83_SimpleRecord/src/main/java/modern/challenge/MelonRecord.java
@@ -1,3 +1,3 @@
 package modern.challenge;
 
-public record MelonRecord(String type, int weight) {}
+public record MelonRecord(String type, float weight) {}
